requirements.txt
scikit-learn
python-slugify
xgboost
matplotlib
scipy
sklego
lightgbm
catboost
tensorflow
tabularize
seaborn
black
flake8
tabulate
